// apps/tourism/platform/src/types/hotel.ts
export interface definitions {
// ...
  hotelReviewRateTypes: {
    /** Format: double */
    rate?: number;
    type?: definitions["hotelReviewsRate"];
  };
// ...
}